Use fixed release of googletest
Recently the master branch of the googletest repo was renamed to main.
Due to this clone of the repository with treep failed (as it assumes
master by default).

Since it might anyway be better to use a fixed release instead of the
(potentially unstable) master/main branch specify the tag of the
(currently) latest release.

Since by default treep first clones the "master" branch and then checks
out the tag, it is necessary to additionally specify `branch: main`.