Add ament_googletest
This is more or less a duplicate of the original googletest repo
(already included in the config) but with some modifications to be used
by ament.

Unfortunately, we need this for 'ament_cmake_gmock' to work while we
also still need to keep the vanilla googletest for packages that don't
use ament.  At least the presence of both in the same workspace does not
seem to cause any trouble.